Truncating part 2 and adding Powershell warning
Truncated the instructions in part 2 to account for the fact that leases collection is now created automatically in ChangeFeedProcessor.cs. This needed to be changed because creating a fixed collection from the portal (as instructed originally) has been disallowed. Also added a warning about possible Powershell errors when running on the latest version of Powershell.